Move the metabot slash command slot into metabot
